WIP: #149 refactoring. Tuple operators extracted.
