+++ akumulid/CMakeLists.txt
@@ -41,9 +41,9 @@ set(CPACK_DEBIAN_PACKAGE_MAINTAINER "Evgeny Lazin")
 set(CPACK_DEBIAN_PACKAGE_DESCRIPTION "Time-series storage engine")
 set(CPACK_DEBIAN_PACKAGE_SECTION "devel")
 
-set(CPACK_PACKAGE_VERSION "0.4.0")
+set(CPACK_PACKAGE_VERSION "0.7.0")
 set(CPACK_PACKAGE_VERSION_MAJOR "0")
-set(CPACK_PACKAGE_VERSION_MINOR "4")
+set(CPACK_PACKAGE_VERSION_MINOR "7")
 set(CPACK_PACKAGE_VERSION_PATCH "0")
 
 include(CPack)
