+++ unittests/CMakeLists.txt
@@ -187,6 +187,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/operator.cpp
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
+    ../libakumuli/storage_engine/operators/join.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/query_processing/queryparser.cpp
     # query processor
@@ -342,6 +343,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/operator.cpp
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
+    ../libakumuli/storage_engine/operators/join.cpp
     ../libakumuli/util.cpp
     ../libakumuli/status_util.cpp
     ../libakumuli/log_iface.cpp
@@ -370,6 +372,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/operator.cpp
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
+    ../libakumuli/storage_engine/operators/join.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/util.cpp
     ../libakumuli/status_util.cpp
